scratchpad/stream protocol move
