+++ libs/langgraph/langgraph/graph/_node.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 from collections.abc import Sequence
 from typing import Any, NamedTuple, Protocol, Union
