+++ libs/langgraph/langgraph/managed/base.py
@@ -8,7 +8,7 @@ from typing import (
 
 from typing_extensions import TypeGuard
 
-from langgraph.types import PregelScratchpad
+from langgraph.pregel._scratchpad import PregelScratchpad
 
 V = TypeVar("V")
 U = TypeVar("U")
