+++ libs/langgraph/langgraph/managed/is_last_step.py
@@ -1,7 +1,7 @@
 from typing import Annotated
 
 from langgraph.managed.base import ManagedValue
-from langgraph.types import PregelScratchpad
+from langgraph.pregel._scratchpad import PregelScratchpad
 
 __all__ = ("IsLastStep", "RemainingStepsManager")
