+++ libs/langgraph/langgraph/pregel/_loop.py
@@ -106,18 +106,17 @@ from langgraph.pregel._io import (
     read_channels,
 )
 from langgraph.pregel._read import PregelNode
+from langgraph.pregel._scratchpad import PregelScratchpad
 from langgraph.pregel._utils import get_new_channel_versions, is_xxh3_128_hexdigest
+from langgraph.pregel.protocol import StreamChunk, StreamProtocol
 from langgraph.store.base import BaseStore
 from langgraph.types import (
     All,
     CachePolicy,
     Command,
     PregelExecutableTask,
-    PregelScratchpad,
     RetryPolicy,
-    StreamChunk,
     StreamMode,
-    StreamProtocol,
 )
 
 V = TypeVar("V")
